REFACTORED : Feet value given in double

+++ src/main/java/com/quantitymeasurement/QuantityMeasurement.java
@@ -1,7 +1,7 @@
 package com.quantitymeasurement;
 
 public class QuantityMeasurement {
-    public int feetToInch(int feet) {
+    public double feetToInch(double feet) {
         return feet * 12;
     }
 }
